Add class metadata for new extension.
Added MetadataType attribute to Orders class, and added OrdersMetadata class. Metadata classes are used when you cannot add attributes directly to a class, or when doing so isn't feasible - EF-generated classes for instance will auto-generate anytime the model is updated from the database, losing the data annotations/attributes already on the class.